UI improvements for typing indicators, message list, and image previews
