+++ apps/web/src/styles/modules/MessageBox.module.css
@@ -9,8 +9,12 @@
   display: flex;
   align-items: center;
   transition: border-color 100ms;
-  position: relative;
   z-index: 20;
+  position: absolute;
+  bottom: 0;
+  right: 0.5rem;
+  left: 0;
+  min-height: 100%;
 
   textarea {
     flex: 1;
@@ -34,14 +38,33 @@
 }
 
 .messageBox::before {
+  box-sizing: border-box;
   content: "";
   background: linear-gradient(to top, var(--clr-surface), transparent);
   position: absolute;
+  z-index: -1;
   top: calc(-2.5rem - 2px);
+  left: 0;
   right: 0;
-  left: -2.5rem;
   height: 2.5rem;
-  z-index: 15;
+  transition:
+    top 200ms,
+    height 200ms;
+}
+
+.messageBox.isTyping::before {
+  top: calc(-4.5rem - 2px);
+  height: 4.5rem;
+}
+
+.typingIndicator {
+  position: absolute;
+  top: -24px;
+  left: 10px;
+  color: var(--clr-surface-50);
+  font-size: 14px;
+  font-style: italic;
+  pointer-events: none;
 }
 
 .sendIcon {
@@ -72,7 +95,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  margin: 0 0.25rem 0 0.5rem;
+  margin: 0 1rem;
 }
 
 .imageButton:hover {
@@ -89,16 +112,16 @@
   display: flex;
   flex-direction: column;
   gap: 0.35rem;
-  margin-bottom: 0.4rem;
+  margin: 0.3rem 0;
 }
 
 .preview {
   display: flex;
   align-items: center;
   gap: 0.5rem;
-  background: var(--clr-dark);
-  border-radius: var(--radius-md);
-  padding: 0.4rem;
+  background: var(--clr-surface-20);
+  border-radius: var(--radius-sm);
+  padding-right: 0.5rem;
 }
 
 .previewImage {
