+++ apps/web/src/styles/modules/MessageListRenderer.module.css
@@ -1,8 +1,6 @@
 .messageListRenderer {
-  margin: 0 1rem;
-  margin-left: 0.5rem;
+  margin-right: 0.5rem;
   display: flex;
-  width: 100%;
   height: calc(100vh - 6rem);
   flex-direction: column-reverse;
   row-gap: 0.1rem;
@@ -12,10 +10,4 @@
   z-index: 0;
   justify-items: flex-end;
   scrollbar-width: none;
-}
-.typingIndicator {
-  color: #9aa6c4;
-  font-size: 14px;
-  font-style: italic;
-  margin-bottom: 8px;
 }
